浅間二丁目のデータ取得を確認
